// src/app/enum/action-type.ts
export enum ActionType {
    DECLARE_INFLUENCE = 'DECLARE_INFLUENCE',
    SEND_HOSPITAL = 'SEND_HOSPITAL',
    EXIT_HOSPITAL = 'EXIT_HOSPITAL',
    PURGE_ATTEMPT = 'PURGE_ATTEMPT',
    EXILE_ESCAPE = 'EXILE_ESCAPE',
    EXILE_RETURN = 'EXILE_RETURN',
    BEGIN_INVESTIGATION = 'BEGIN_INVESTIGATION',
    REMOVE_INVESTIGATION = 'REMOVE_INVESTIGATION',
    OPEN_TRIAL = 'OPEN_TRIAL',
    CAST_TRIAL_VOTE = 'CAST_TRIAL_VOTE',
    CONDEMNATION = 'CONDEMNATION',
    NEGATE_CONDEMNATION = 'NEGATE_CONDEMNATION'
}